:recycle: (drivers): Code Smells

+++ include/interface/drivers/DacTimer.h
@@ -14,15 +14,15 @@ class DacTimer
   public:
 	~DacTimer() = default;
 
-	virtual void initialize(float frequency) = 0;
-	virtual void deInitialize()				 = 0;
-	virtual void start()					 = 0;
-	virtual void stop()						 = 0;
+	virtual void initialize(uint32_t frequency) = 0;
+	virtual void deInitialize()					= 0;
+	virtual void start()						= 0;
+	virtual void stop()							= 0;
 
 	virtual auto getHandle() -> TIM_HandleTypeDef = 0;
 
   protected:
-	virtual auto _calculatePeriod(float freq) -> uint32_t = 0;
+	virtual auto _calculatePeriod(uint32_t freq) -> uint32_t = 0;
 
 	virtual void _registerMspCallbacks() = 0;
 	virtual void _mspInitCallback()		 = 0;
